+++ src/renderer/App.tsx
@@ -18,15 +18,24 @@ import BottomStatusBar from './BottomStatusBar';
 import './App.css';
 import { useSettings } from './useSettings';
 import { getCategoryFromConfig } from './rendererutils';
+import StateManager from './StateManager';
 
 const ipc = window.electron.ipcRenderer;
 
 const WarcraftRecorder = () => {
   const [config] = useSettings();
   const [error, setError] = useState<string>('');
-  const [videoState, setVideoState] = useState<RendererVideo[]>([]);
   const [micStatus, setMicStatus] = useState<MicStatus>(MicStatus.NONE);
   const [crashes, setCrashes] = useState<Crashes>([]);
+
+  // The video state contains most of the frontend state, it's complex so
+  // frontend triggered modifications go through the StateManager class, which
+  // calls the React set function appropriately.
+  const [videoState, setVideoState] = useState<RendererVideo[]>([]);
+
+  const stateManager = useRef<StateManager>(
+    StateManager.getInstance(setVideoState)
+  );
 
   const [recorderStatus, setRecorderStatus] = useState<RecStatus>(
     RecStatus.WaitingForWoW
@@ -63,9 +72,11 @@ const WarcraftRecorder = () => {
   // Used to allow for hot switching of video players when moving between POVs.
   const persistentProgress = useRef(0);
 
+  // Used to remember the player height when switching categories.
+  const playerHeight = useRef(500);
+
   const doRefresh = async () => {
-    const state = (await ipc.invoke('getVideoState', [])) as RendererVideo[];
-    setVideoState(state);
+    stateManager.current.refresh();
 
     setAppState((prevState) => {
       return {
@@ -126,10 +137,12 @@ const WarcraftRecorder = () => {
       <RendererTitleBar />
       <Layout
         recorderStatus={recorderStatus}
+        stateManager={stateManager}
         videoState={videoState}
         appState={appState}
         setAppState={setAppState}
         persistentProgress={persistentProgress}
+        playerHeight={playerHeight}
       />
       <BottomStatusBar
         recorderStatus={recorderStatus}