+++ src/renderer/TagDialog.tsx
@@ -8,12 +8,14 @@ import {
   Button,
 } from '@mui/material';
 import { RendererVideo } from 'main/types';
-import { Dispatch, SetStateAction } from 'react';
+import { Dispatch, MutableRefObject, SetStateAction } from 'react';
+import StateManager from './StateManager';
 
 interface IProps {
   video: RendererVideo;
   tagDialogOpen: boolean;
   setTagDialogOpen: Dispatch<SetStateAction<boolean>>;
+  stateManager: MutableRefObject<StateManager>;
 }
 
 const buttonSx = {
@@ -26,7 +28,7 @@ const buttonSx = {
 };
 
 export default function TagDialog(props: IProps) {
-  const { video, tagDialogOpen, setTagDialogOpen } = props;
+  const { video, tagDialogOpen, setTagDialogOpen, stateManager } = props;
 
   const closeTagDialog = (event: React.MouseEvent<HTMLElement>) => {
     event.stopPropagation();
@@ -34,6 +36,8 @@ export default function TagDialog(props: IProps) {
   };
 
   const saveTag = (newTag: string) => {
+    stateManager.current.tag(video, newTag);
+
     window.electron.ipcRenderer.sendMessage('videoButton', [
       'tag',
       video.videoSource,